Support prefix-matches
