+++ LICENSE
@@ -1,4 +1,4 @@
-Copyright 2019 Black Square Media Ltd
+Copyright 2020 Black Square Media Ltd
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
