+++ spec/searchable_by_spec.rb
@@ -1,38 +1,6 @@
 require 'spec_helper'
 
-describe ActiveRecord::SearchableBy do
-  context 'norm_values' do
-    def norm(str)
-      described_class.norm_values(str).each_with_object({}) do |val, acc|
-        acc[val.term] = val.negate
-      end
-    end
-
-    it 'should tokenise strings' do
-      expect(norm(nil)).to eq({})
-      expect(norm('""')).to eq({})
-      expect(norm('-+""')).to eq({})
-      expect(norm('simple words')).to eq('simple' => false, 'words' => false)
-      expect(norm(" with   \t spaces\n")).to eq('with' => false, 'spaces' => false)
-      expect(norm('with with duplicates with')).to eq('with' => false, 'duplicates' => false)
-      expect(norm('with "full term"')).to eq('full term' => false, 'with' => false)
-      expect(norm('"""odd double quotes around"""')).to eq('odd double quotes around' => false)
-      expect(norm('""even double quotes around""')).to eq('even double quotes around'=> false)
-      expect(norm('with\'apostrophe')).to eq("with'apostrophe" => false)
-      expect(norm('with -minus')).to eq('minus' => true, 'with' => false)
-      expect(norm('with +plus')).to eq('plus' => false, 'with' => false)
-      expect(norm('with-minus')).to eq('with-minus' => false)
-      expect(norm('with+plus')).to eq('with+plus' => false)
-      expect(norm('with -"minus before"')).to eq('minus before' => true, 'with' => false)
-      expect(norm('with "-minus within"')).to eq('-minus within' => false, 'with' => false)
-      expect(norm('with +"plus before"')).to eq('plus before' => false, 'with' => false)
-      expect(norm('with "+plus within"')).to eq('+plus within' => false, 'with' => false)
-      expect(norm('+plus "in other term"')).to eq('in other term' => false, 'plus' => false)
-      expect(norm('with_blank \'\'')).to eq('with_blank' => false, '\'\'' => false)
-      expect(norm('with_blank_doubles ""')).to eq('with_blank_doubles' => false)
-    end
-  end
-
+describe SearchableBy do
   it 'should ignore bad inputs' do
     expect(Post.search_by(nil).count).to eq(5)
     expect(Post.search_by('').count).to eq(5)
@@ -67,6 +35,17 @@ describe ActiveRecord::SearchableBy do
     expect(Post.search_by('+alice "your recipe"').pluck(:title)).to match_array(%w[a2])
     expect(Post.search_by('bob -"her recipe"').pluck(:title)).to match_array(%w[b2 ab])
     expect(Post.search_by('bob +"her recipe"').pluck(:title)).to match_array(%w[b1])
+  end
+
+  it 'should respect match options' do
+    # name uses match: :prefix
+    expect(Post.search_by('alice').pluck(:title)).to match_array(%w[a1 a2 ab])
+    expect(Post.search_by('ali').pluck(:title)).to match_array(%w[a1 a2 ab])
+    expect(Post.search_by('lice').pluck(:title)).to be_empty
+    expect(Post.search_by('li').pluck(:title)).to be_empty
+
+    # title uses match: :all (default)
+    expect(Post.search_by('recip').pluck(:title)).to match_array(%w[a1 a2 b1 b2 ab])
   end
 
   it 'should search within scopes' do
